+++ .eslintrc.js
@@ -1,20 +1,24 @@
 module.exports = {
+    root: true,
     parser: '@typescript-eslint/parser',
     parserOptions: {
         ecmaVersion: 2020,
         sourceType: 'module',
     },
+    plugins: ['@typescript-eslint'],
     extends: [
         'eslint:recommended',
+        'plugin:@typescript-eslint/recommended'
     ],
     env: {
         node: true,
-        es6: true,
+        es2021: true,
+        jest: true,
     },
     rules: {
         'no-console': 'off',
-        'no-unused-vars': 'off',
-        '@typescript-eslint/no-unused-vars': 'off',
+        '@typescript-eslint/no-unused-vars': ['warn'],
+        '@typescript-eslint/no-explicit-any': 'off',
     },
+    ignorePatterns: ['dist/', 'node_modules/'],
 };
-